# packages/server/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_mhmadamrii_kommers_server_internal_model.OrderStatus:
    enum:
    - pending
    - paid
    - cancelled
    type: string
    x-enum-varnames:
    - OrderStatusPending
    - OrderStatusPaid
    - OrderStatusCancelled
  github_com_mhmadamrii_kommers_server_internal_model.PaymentStatus:
    enum:
    - pending
    - paid
    - failed
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusPaid
    - PaymentStatusFailed
...
  internal_handler.checkoutRequest:
    properties:
      address_id:
        type: integer
    required:
    - address_id
    type: object
...
  internal_handler.orderItemResponse:
    properties:
      campaign_id:
        type: integer
      id:
        type: integer
      price_cents:
        type: integer
      product_id:
        type: integer
      product_name:
        type: string
      quantity:
        type: integer
      subtotal_cents:
        type: integer
    type: object
  internal_handler.orderResponse:
    properties:
      address_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal_handler.orderItemResponse'
        type: array
      payment_status:
        $ref: '#/definitions/github_com_mhmadamrii_kommers_server_internal_model.PaymentStatus'
      status:
        $ref: '#/definitions/github_com_mhmadamrii_kommers_server_internal_model.OrderStatus'
      total_cents:
        type: integer
    type: object
...
info:
  contact: {}
  description: E-commerce backend API for kommers.
  title: Kommers API
  version: "1.0"
paths:
...
  /api/v1/checkout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Checkout payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.checkoutRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handler.orderResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Convert the current cart into an order (payment integration pending)
      tags:
      - orders
...
  /api/v1/orders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handler.orderResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List the current user's orders
      tags:
      - orders
  /api/v1/orders/{id}:
    get:
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.orderResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get an order by ID
      tags:
      - orders
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"